+++ crates/common/src/registration.rs
@@ -1,5 +1,4 @@
-use crate::Value;
-use serde::{Deserialize, Serialize};
+use crate::{Deserialize, Serialize, Value};
 use std::collections::HashMap;
 
 /// The `Registration` type is used by plugins in the `init` function to expose
@@ -7,29 +6,33 @@ use std::collections::HashMap;
 /// runtime.
 #[derive(Debug, Default, Serialize, Deserialize)]
 pub struct Registration {
+    #[serde(rename = "n")]
     pub name: String,
-    pub write: HashMap<String, Value>,
-    pub read: HashMap<String, Vec<String>>,
+    #[serde(rename = "s")]
+    pub state: Option<HashMap<String, Value>>,
+    #[serde(rename = "d")]
+    pub dependencies: Option<Vec<String>>,
 }
 
 impl Registration {
     pub fn new(name: impl Into<String>) -> Self {
         Registration {
             name: name.into(),
-            write: HashMap::default(),
-            read: HashMap::default(),
+            ..Self::default()
         }
     }
 
-    pub fn write(mut self, key: impl Into<String>, value: Value) -> Self {
-        self.write.insert(key.into(), value);
+    /// Define a key/value pair of state this plugin wants to manage.
+    pub fn state(mut self, key: impl Into<String>, value: Value) -> Self {
+        self.state
+            .get_or_insert(HashMap::default())
+            .insert(key.into(), value);
+
         self
     }
 
-    pub fn read(mut self, plugin: impl Into<String>, keys: &[&'static str]) -> Self {
-        let keys = keys.to_vec().into_iter().map(Into::into).collect();
-
-        self.read.insert(plugin.into(), keys);
+    pub fn dependency(mut self, name: impl Into<String>) -> Self {
+        self.dependencies.get_or_insert(vec![]).push(name.into());
         self
     }
 }