+++ crates/sdk/src/lib.rs
@@ -2,6 +2,8 @@ mod error;
 mod internal;
 mod load;
 pub mod prelude;
+mod sdk;
 
 pub use error::Error;
 pub use internal::{init, malloc, run};
+pub use sdk::Sdk;